Clear the cache, delete files, do not use clean folders, so keep the folder

+++ DLClent_A/src/com/darly/dlclent/common/CleanCache.java
@@ -58,7 +58,9 @@ public class CleanCache {
 				}
 			}
 		}
-		dir.delete();
+		if (dir.isFile()) {
+			dir.delete();
+		}
 		// 目录此时为空，可以删除
 		return true;
 	}
